update profile.db, page.tsx, route.js and tsconfig.json

+++ src/app/api/post/[id]/route.js
@@ -1,7 +1,7 @@
 import { apiDelete, apiGet } from "@/app/api/database";
 import { NextResponse } from "next/server";
 
-export async function DELETE(request) {
+export async function GET(request) {
   try {
     const url = new URL(request.url);
     const id = url.pathname.split("/").pop();
@@ -12,37 +12,46 @@ export async function DELETE(request) {
         { status: 400 }
       );
 
-    const query = "DELETE FROM blogposts WHERE id = ?";
+    const query = `SELECT * FROM blogposts WHERE id = ${id}`;
+    const post = await apiGet(query);
 
-    const result = await apiDelete(query, [id]);
-
-    if (result.changes === 0)
+    if (!post)
       return NextResponse.json({ error: "Post not found" }, { status: 404 });
 
-    return NextResponse.json({ message: "Post deleted successfully" });
+    return NextResponse.json(post[0]);
   } catch (error) {
-    console.error("Failed to delete post:", error);
+    console.error(`Failed to fetch post: ${error.message}`, error);
     return NextResponse.json(
-      { error: "Failed to delete post" },
+      { error: "Failed to fetch post" },
       { status: 500 }
     );
   }
 }
 
-export async function GET(request) {
+export async function DELETE(request) {
   try {
-    const { searchParams } = new URL(request.url);
-    const id = searchParams.get("id");
+    const url = new URL(request.url);
+    const id = url.pathname.split("/").pop();
 
-    let query = `SELECT * FROM blogposts WHERE id = '${id}`;
+    if (!id) {
+      return NextResponse.json(
+        { error: "Post ID is required" },
+        { status: 400 }
+      );
+    }
 
-    const post = await apiGet(query);
-    console.log("hey", post);
-    return NextResponse.json(post);
+    const query = "DELETE FROM blogposts WHERE id = ?";
+    const result = await apiDelete(query, [id]);
+
+    if (result.changes === 0) {
+      return NextResponse.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    return NextResponse.json({ message: "Post deleted successfully" });
   } catch (error) {
-    console.error("Failed to fetch post:", error);
+    console.error(`Failed to delete post: ${error.message}`, error);
     return NextResponse.json(
-      { error: "Failed to fetch post" },
+      { error: "Failed to delete post" },
       { status: 500 }
     );
   }
